fix: restructure component imports and update package names
