+++ design_system/src/main/res/values-ar/strings.xml
@@ -37,7 +37,6 @@
     <string name="an_error_occured_while_getting_results">هناك خطأ حدث عند جلب نتائج البحث</string>
     <string name="an_unexpected_error_occurred">خطأ غير متوقع حدث</string>
     <string name="an_unexpected_error_occurred_description">هناك خطأ حدث في تنفيث طلبك. قم بتفقد اتصالك وحاول مرة اخرى</string>
-
     <string name="for_you">من أجلك</string>
     <string name="see_all">عرض الكل</string>
     <string name="explore_more">اكتشف المزيد</string>
