+++ ui/src/main/res/values-ar/strings.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="icon">ايقونة</string>
     <string name="reviewer_image">صورة المراجع</string>
     <string name="rating_star">نجمة التقييم</string>
     <string name="rate">قّيم</string>
