+++ ui/src/main/res/values/strings.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="explore_more">Explore more</string>
-    <string name="icon">Icon</string>
     <string name="reviewer_image">Reviewer Image</string>
     <string name="rating_star">Rating Star</string>
     <string name="rate">Rate it</string>
